MCC-1204 Complete process contracts fees for insurance

+++ app/Http/Casts/InsurancePlan/ContractFeesRequestCast.php
@@ -61,10 +61,10 @@ final class ContractFeesRequestCast extends CastsRequest
         return collect($this->inputs['modifier_ids'] ?? []);
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return array_key_exists('price', $this->inputs)
-            ? (int) $this->inputs['price']
+            ? (float) $this->inputs['price']
             : null;
     }
 
@@ -110,10 +110,10 @@ final class ContractFeesRequestCast extends CastsRequest
             : null;
     }
 
-    public function getPricePercentage(): ?int
+    public function getPricePercentage(): ?float
     {
         return array_key_exists('price_percentage', $this->inputs)
-            ? (int) $this->inputs['price_percentage']
+            ? (float) $this->inputs['price_percentage']
             : null;
     }
